Add files via upload
Suppose you want to generate a table for a function f(i,j), for 1 ≤ i ≤ N and 1 ≤j ≤ M. This can be accomplished easily using a double for loop. In a functional programming language like R, the emphasis is to use in-built functions that can perform the task, instead of using a for loop. In this article, I will illustrate how the sapply() function can be used to get the job done using few lines of code.